fix: use radio button for software types filter
