+++ assets/js/components/Catalogue/CatalogueView.js
@@ -4,13 +4,12 @@ import { CatalogueSummary } from './CatalogueSummary.js';
 import { CatalogueItems } from './CatalogueItems.js';
 import { useSearchEngine } from '../../hooks/useSearchEngine.js';
 import { searchContextState } from '../../contexts/searchContext.js';
-import { SOFTWARE_OPEN, SOFTWARE_REUSE } from '../../utils/constants.js';
 import { l10NLabels } from '../../utils/l10n.js';
 
 export const CatalogueView = React.memo(() => {
   const { filterCategories, filterDevelopmentStatuses, filterIntendedAudiences, type } = useContext(searchContextState);
   let totalAppliedFilters = filterCategories.length + filterIntendedAudiences.length + filterDevelopmentStatuses.length;
-  if ([SOFTWARE_OPEN, SOFTWARE_REUSE].includes(type)) {
+  if (type) {
     totalAppliedFilters++;
   }
